volumes/app/requirements.txt
asgiref==3.5.2
Django==3.2.15
django-filter==22.1
django-taggit==3.0.0
djangorestframework==3.13.1
Markdown==3.4.1
mysqlclient==2.1.1
Pillow==9.2.0
python-decouple==3.6
python-slugify==6.1.2
pytz==2022.6
slugify==0.0.1
sqlparse==0.4.3
text-unidecode==1.3
celery==5.2.7
django_redis
pytest